Prefs dialog works, need to add new code from master and refactor
